fix: Проводит модуляцию js файлов
